requirements.txt
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
Pillow==9.1.0
scikit_learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
streamlit==1.8.1
